refactor: вынесла функуию getBase64 в папку libs
