+++ packages/client/src/shared/ui/FileInput/libs/index.ts
@@ -1,0 +1,2 @@
+export { getBase64 } from './getBase64'
+export type { FileType as type } from './getBase64'
